requirements.txt
fastapi
pika
cassandra-driver